requirements.txt
Flask==0.10.1
Flask-API==0.6.2
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==1.6
RPi.GPIO==0.5.7
Werkzeug==0.9.6
audioread==1.0.3
bpython==0.13.1
itsdangerous==0.24
librosa==0.3.0
matplotlib==1.4.2
mock==1.0.1
nose==1.3.4
numpy==1.9.0
pyparsing==2.0.3
python-dateutil==2.2
pytz==2014.7
scikit-learn==0.15.2
scipy==0.14.0
six==1.8.0
wsgiref==0.1.2
